Move IDX parts to membership page
